fixed Vectorized Tiling

+++ python/test_matmul.py
@@ -152,8 +152,8 @@ def test_matmul():
     cols1 = 167
     rows2 = 167
     cols2 = 321
-    a = torch.randn((rows1, cols1), device='cpu', dtype=torch.float32)
-    b = torch.randn((rows2, cols2), device='cpu', dtype=torch.float32)
+    a = torch.randn((rows1, cols1), device='cpu', dtype=torch.float16)
+    b = torch.randn((rows2, cols2), device='cpu', dtype=torch.float16)
     # a = torch.full((rows1, cols1), 1, device='cpu', dtype=torch.float32)
     # b = torch.full((rows2, cols2), 1, device='cpu', dtype=torch.float32)
     triton_output = matmul(a, b)
